Browser support
The application is now styled for support in Chrome, Opera, Safari, and
Firefox. Edge seems to be styled properly, but has some sort of
Javascript error that requires attention for it to be functional. We
will not be support IE, period. At least until someone decides to pitch
in on a custom stylesheet for it (should that ever happen).